Fix empty database name in local dev environment
